+++ tests/web_util_tests.py
@@ -39,13 +39,12 @@ class WebUtilTests(unittest.TestCase):
         try:
             links = scrape_page_for_links("neon.ai")
             self.assertIsInstance(links, dict)
-            self.assertIn("about us", links.keys())
+            self.assertIn("company", links.keys())
             # TODO: Update test to validate absolute and relative URL paths
             # Relative href
-            self.assertIn(links["about us"],
-                          ("https://neon.ai/aboutus",
-                           "http://neon.ai/aboutus",
-                           "https://neon.ai/index.php/aboutus"))
+            self.assertIn(links["company"],
+                          ("https://neon.ai/company",
+                           "https://neon.ai/company/"))
         except ConnectTimeout:
             LOG.error("Github testing breaks here")